functions/requirements.txt
# functions/requirements.txt
firebase-functions
google-cloud-speech
google-cloud-storage
